// src/bot/Client/EntityTypes.ts
export enum EntityTypeFromString {
  POTION_TIMER_5 = '5',
  POTION_TIMER_4 = '4',
  POTION_TIMER_3 = '3',
  POTION_TIMER_2 = '2',
  POTION_TIMER_1 = '1',
  BLAST = '҉',
  WALL = '☼',
  TREASURE_BOX = '#',
  TREASURE_BOX_OPENING = 'H',
  GHOST = '&',
  GHOST_DEAD = 'x',
  POTION_BLAST_RADIUS_INCREASE = '+',
  POTION_COUNT_INCREASE = 'c',
  POTION_REMOTE_CONTROL = 'r',
  POTION_IMMUNE = 'i',
  POISON_THROWER = 'T',
  POTION_EXPLODER = 'A',
  NONE = ' ',
  HERO = '☺',
  HERO_POTION = '☻',
  HERO_DEAD = 'Ѡ',
  OTHER_HERO = '♥',
  OTHER_HERO_POTION = '♠',
  OTHER_HERO_DEAD = '♣',
  ENEMY_HERO = 'ö',
  ENEMY_HERO_POTION = 'Ö',
  ENEMY_HERO_DEAD = 'ø',
}

export enum EntityType {
  POTION_TIMER_5 = 'POTION_TIMER_5',
  POTION_TIMER_4 = 'POTION_TIMER_4',
  POTION_TIMER_3 = 'POTION_TIMER_3',
  POTION_TIMER_2 = 'POTION_TIMER_2',
  POTION_TIMER_1 = 'POTION_TIMER_1',
  BLAST = 'BLAST',
  WALL = 'WALL',
  TREASURE_BOX = 'TREASURE_BOX',
  TREASURE_BOX_OPENING = 'TREASURE_BOX_OPENING',
  GHOST = 'GHOST',
  GHOST_DEAD = 'GHOST_DEAD',
  POTION_BLAST_RADIUS_INCREASE = 'POTION_BLAST_RADIUS_INCREASE',
  POTION_COUNT_INCREASE = 'POTION_COUNT_INCREASE',
  POTION_REMOTE_CONTROL = 'POTION_REMOTE_CONTROL',
  POTION_IMMUNE = 'POTION_IMMUNE',
  POISON_THROWER = 'POISON_THROWER',
  POTION_EXPLODER = 'POTION_EXPLODER',
  NONE = 'NONE',
  HERO = 'HERO',
  HERO_POTION = 'HERO_POTION',
  HERO_DEAD = 'HERO_DEAD',
  OTHER_HERO = 'OTHER_HERO',
  OTHER_HERO_POTION = 'OTHER_HERO_POTION',
  OTHER_HERO_DEAD = 'OTHER_HERO_DEAD',
  ENEMY_HERO = 'ENEMY_HERO',
  ENEMY_HERO_POTION = 'ENEMY_HERO_POTION',
  ENEMY_HERO_DEAD = 'ENEMY_HERO_DEAD',
}